fix: Additional Coverity defect fixes (15+ CIDs)
Integer Overflow Prevention:
- CID 1675228: async_broadcaster - Cast to int64_t before shift in retry delay
- CID 1675211: validation.cpp - Cast to uint64_t before shift to prevent sign extension
- CID 1675261: blockencodings - Validate untrusted allocation size
- CID 1675268: blockencodings - Validate transaction count in Deserialize

Performance (Copy Avoidance):
- CID 1675171: Multiple files - Use std::move and move iterators to avoid copies
  - miner/controller.cpp: coinbaseIn, coinbaseOut, selectedTxs, blockTxData
  - wallet/wal.cpp: buffer, entry
  - blockencodings.cpp: txData, result vectors

Thread Safety:
- CID 1675230: miner/controller - Sleep outside lock to prevent blocking
- CID 1675314: net.cpp - Release lock before blocking send, re-verify socket

Unchecked Return Values:
- CID 1675254: rpc/logger.cpp - Check seekp/tellp return values in log rotation
- CID 1675231: rpc/logger.cpp - Check std::rename/std::remove return values
- CID 1675266: wallet.cpp - Check tellp/seekg/seekp/read return values
- CID 1675271: http_server.cpp - Check send() return value
- CID 1675303: websocket.cpp - Check setsockopt return value

Input Validation:
- CID 1675263: wal.cpp - Validate WAL entry count against MAX_WAL_ENTRIES

+++ src/consensus/validation.cpp
@@ -111,15 +111,23 @@ bool CBlockValidator::DeserializeBlockTransactions(
             error = "Incomplete transaction count (253)";
             return false;
         }
-        txCount = data[offset] | (data[offset + 1] << 8);
+        // CID 1675211 FIX: Cast to uint64_t before shifting to prevent sign extension
+        // When shifting uint8_t values, they are promoted to int (signed), which can cause
+        // sign extension if the high bit is set. Casting to uint64_t ensures unsigned behavior.
+        txCount = static_cast<uint64_t>(data[offset]) | (static_cast<uint64_t>(data[offset + 1]) << 8);
         offset += 2;
     } else if (firstByte == 254) {
         if (offset + 4 > dataSize) {
             error = "Incomplete transaction count (254)";
             return false;
         }
-        txCount = data[offset] | (data[offset + 1] << 8) |
-                 (data[offset + 2] << 16) | (data[offset + 3] << 24);
+        // CID 1675211 FIX: Cast to uint64_t before shifting to prevent sign extension
+        // When shifting uint8_t values, they are promoted to int (signed), which can cause
+        // sign extension if the high bit is set. Casting to uint64_t ensures unsigned behavior.
+        txCount = static_cast<uint64_t>(data[offset]) |
+                 (static_cast<uint64_t>(data[offset + 1]) << 8) |
+                 (static_cast<uint64_t>(data[offset + 2]) << 16) |
+                 (static_cast<uint64_t>(data[offset + 3]) << 24);
         offset += 4;
     } else {
         error = "Unsupported transaction count encoding (255)";
